requirements.txt
horovod==0.19.5
pytest==5.4.3
scipy==1.4.1
kaldi_python_io==1.1.3
tqdm==4.42.1
editdistance==0.5.3
librosa==0.7.2
torch_complex==0.2.0
numpy==1.18.1
SoundFile==0.10.2
torch==1.6.0
kenlm==0.0.0
PyYAML==5.3.1
warp_rnnt==0.4.0
tensorboard==2.2.1
pre-commit==2.7.1
yapf==0.30.0
museval==0.3.1
pystoi==0.3.3
pypesq==1.2.4
